+++ src/main/java/ru/chernyshev/recognizer/repository/UserRepository.java
@@ -1,9 +1,8 @@
 package ru.chernyshev.recognizer.repository;
 
 import org.springframework.data.repository.CrudRepository;
-import ru.chernyshev.recognizer.entity.ChatEntity;
 import ru.chernyshev.recognizer.entity.UserEntity;
 
 public interface UserRepository extends CrudRepository<UserEntity, Long> {
-    UserEntity findByTelegramId(Integer telegramId);
+    UserEntity findByTelegramId(Long telegramId);
 }
